requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.5.0
opencv-python>=4.7.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Explainability
grad-cam>=1.4.6
pytorch-grad-cam>=1.4.8

# Deployment (for Streamlit app)
streamlit>=1.28.0
flask>=2.3.0

# Utilities
pyyaml>=6.0
albumentations>=1.3.0